fixed wingbuffet, screenshake, decrementing resistances improved stoneskin, decrementing resistances

+++ gemrb/core/GlobalTimer.h
@@ -45,7 +45,7 @@ private:
 	int fadeFromCounter, fadeFromMax;
 	unsigned long waitCounter;
 	int shakeCounter;
-	unsigned long shakeX, shakeY;
+	int shakeX, shakeY;
 	unsigned int first_animation;
 	std::vector<AnimationRef*>  animations;
 	//move viewport to this coordinate
@@ -67,8 +67,7 @@ public:
 	void SetFadeToColor(unsigned long Count);
 	void SetFadeFromColor(unsigned long Count);
 	void SetWait(unsigned long Count);
-	void SetScreenShake(unsigned long shakeX, unsigned long shakeY,
-		unsigned long Count);
+	void SetScreenShake(int shakeX, int shakeY, unsigned long Count);
 	void AddAnimation(ControlAnimation* ctlanim, unsigned long time);
 	void RemoveAnimation(ControlAnimation* ctlanim);
 	void ClearAnimations();
